Add testing infrastructure (vitest/rtl/msw)
We'll use a high-level UI testing approach to ensure that our app is
working as expected.

Vitest serves as our test runner / framework (including bells &
whistles, like a UI tester and code coverage reporting).

React Testing Library (rtl) with a JSDOM allows us to conveniently
render our app and interact with it. It's not full UI testing in the
browser, but very close to it.

Finally, we throw in Mock Service Worker (msw) to be able to mock the
API calls made by the app. Obviously we don't want to make real requests
to the GitHub APIs when running tests (both for performance and
reliability reasons).