Introduce React Query for API requests
TanStack Query and its React adapter provide a powerful way to deal with
asynchronous data and state management.

The GitHub API request is moved out from the Repos component, and
re-implemented using React Query in data/github.ts. This way, we have
separation of concerns: the data module deals with the API call and
loading/error handling, and the component module is using it via a
custom hook and only deals with UI concerns, like showing a loading
indicator or the loaded data.

The tests are slightly adapted, but only the setup part. React Query is
by default doing caching and retrying on API failures. We could keep
that behaviour and modify our tests, but it would make them more
complicated without any benefit. Therefore, the retry mechanism is
disabled when the App is bootstrapped in the tests.

See: https://tanstack.com/query/v4