Calculate date of one week ago and use in query
The GitHub API call was still using a hardcoded date. We actually want
to retrieve repositories created in the last week, so we'll have to pass
the proper date string as a query parameter.

Date handling in JS is always a pain. There are libraries like date-fns
and momentjs to help with this, but as we are only having to deal with
one single date calculation here, we can go with our own implementation.

To be sure that everything is working fine, even in edge cases like
passing a month or year, an appropriate amount of units are added.