Split Repo component
As the Repo component was both dealing with the data fetching and the
rendering, it makes sense to split that up. Only things related to the
data loading (loading and error indicators) remain in the old component
(now called RepositoryWrapper).

When the repository data is loaded successful, it is passed down to a
new Repositories component.

We'll soon show more details about the repositories and add some nice
layout, which can then be located in that component.