Toggle favorites with buttons, persist in storage
Clicking the favorite button on a repository card now actually toggles
the state and marks the respective repo as a favorite. Clicking once
more removes the favorite (indicated by a changing button text).

Those states are put into local storage (which is read initially for
initialization), therefore providing persistence on the browser level.