Show number of saved favorites in toggle button
It's nice to be able to see how many favorites are already saved. This
information can then also be used to disable the "only favorites" button
when there aren't any.

To be able to do this, the favorites state has to be shared between the
Repositories and the FavoritesFilterButton components. To do that, the
useLocalStorage hook usage is moved up in the component tree to the App
component, and the state / state setter function are passed down as
props.